shared error
